backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.3.0
qdrant-client==1.7.0
pypdf2==3.0.1
pdfplumber==0.10.3
sentence-transformers==2.2.2
numpy==1.24.3
python-multipart==0.0.6
aiofiles==23.2.1
gunicorn==21.2.0

